Organize ReelMagic as a sub-module under hardware
src/hardware/reelmagic/driver.cpp
src/hardware/reelmagic/mpeg_decoder.h
src/hardware/reelmagic/player.cpp
src/hardware/reelmagic/vga_passthrough.h
src/hardware/reelmagic/video_mixer.cpp

src/hardware/reelmagic/docs/...
src/hardware/reelmagic/tools/...

This matches the other multi-component modules that have
their own subdirectory like: mame, mouse, and serialport.

This allows for dropping the redundant filename prefixes, and
simply call the component-code was they represent within the
Reel Magic design.

It also co-locates the tools and documentation, and makes
small modifications to reflect the new paths and names.
