+++ src/hardware/meson.build
@@ -48,9 +48,9 @@ libhardware_sources = files(
     'pcspeaker_impulse.cpp',
     'pic.cpp',
     'ps1audio.cpp',
-    'reelmagic_driver.cpp',
-    'reelmagic_player.cpp',
-    'reelmagic_videomixer.cpp',
+    'reelmagic/driver.cpp',
+    'reelmagic/player.cpp',
+    'reelmagic/video_mixer.cpp',
     'sblaster.cpp',
     'ston1_dac.cpp',
     'tandy_sound.cpp',
