+++ src/hardware/reelmagic/docs/hardware_io_ports.md
