+++ src/hardware/reelmagic/docs/known_game_bugs.md
