+++ src/hardware/reelmagic/tools/wincompat.h
