+++ vs/dosbox.vcxproj
@@ -607,9 +607,9 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClCompile Include="..\src\hardware\pcspeaker_impulse.cpp" />
     <ClCompile Include="..\src\hardware\pic.cpp" />
     <ClCompile Include="..\src\hardware\ps1audio.cpp" />
-    <ClCompile Include="..\src\hardware\reelmagic_driver.cpp" />
-    <ClCompile Include="..\src\hardware\reelmagic_player.cpp" />
-    <ClCompile Include="..\src\hardware\reelmagic_videomixer.cpp" />
+    <ClCompile Include="..\src\hardware\reelmagic\driver.cpp" />
+    <ClCompile Include="..\src\hardware\reelmagic\player.cpp" />
+    <ClCompile Include="..\src\hardware\reelmagic\video_mixer.cpp" />
     <ClCompile Include="..\src\hardware\sblaster.cpp" />
     <ClCompile Include="..\src\hardware\serialport\directserial.cpp" />
     <ClCompile Include="..\src\hardware\serialport\libserial.cpp" />
